requirements.txt
python-telegram-bot[job-queue]==21.6
yt-dlp>=2024.10.7
python-dateutil>=2.9.0
